feat: implement Performance monitor
