--- metrics/metrics_test.go
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2019 QLC Chain Team
- *
- * This software is released under the MIT License.
- * https://opensource.org/licenses/MIT
- */
-
-package metrics
-
-import (
-	"testing"
-)
-
-func TestMetrics(t *testing.T) {
-	//go influxdb.InfluxDB(
-	//	metrics.DefaultRegistry, // metrics registry
-	//	time.Second*10,          // interval
-	//	"http://localhost:8086", // the InfluxDB url
-	//	"mydb",                  // your InfluxDB database
-	//	"myuser",                // your InfluxDB user
-	//	"mypassword",            // your InfluxDB password
-	//)
-}
